refactor: Move files into new `prefs` directory
