+++ src/content/services/ui-manager.ts
@@ -91,7 +91,7 @@ export default class UIManager implements Service {
 
   private openPreferences() {
     this.window.openDialog(
-      'chrome://notero/content/preferences.xul',
+      'chrome://notero/content/prefs/preferences.xul',
       'notero-preferences'
     );
   }
